models/requirements.txt
# Requirements for model export and training

# PyTorch
torch>=2.0.0
torchvision>=0.15.0

# YOLO
ultralytics>=8.0.0

# ONNX
onnx>=1.14.0
onnxruntime-gpu>=1.16.0
onnxsim>=0.4.33

# Utilities
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Optional: Training
# albumentations>=1.3.0
# wandb>=0.15.0
